fixed bug on deleting the list

+++ src/engine/dccnit.h
@@ -373,7 +373,7 @@ class engine
       double actualFPS;            /**< Actual FPS */
 
       Uint32 lastKeyb;             /**< Last time read from keyboard */
-      Uint8  lastKey;              /**< Last key with refresh readed */
+      Uint32  lastKey;             /**< Last key with refresh readed */
 
       sound* snd;                  /**< Actual Sounds thing */
       sndfx* walkSound;            /**< Actual walking sound */
